add-separate-logger-class
