+++ src/ansys/additive/core/__init__.py
@@ -47,6 +47,7 @@ if not os.path.exists(EXAMPLES_PATH):  # pragma: no cover
 
 from ansys.additive.core.additive import Additive
 from ansys.additive.core.geometry_file import BuildFile, MachineType, StlFile
+from ansys.additive.core.logger import LOG
 from ansys.additive.core.machine import AdditiveMachine, MachineConstants
 from ansys.additive.core.material import (
     AdditiveMaterial,
@@ -60,7 +61,6 @@ from ansys.additive.core.microstructure import (
     MicrostructureSummary,
 )
 from ansys.additive.core.porosity import PorosityInput, PorositySummary
-from ansys.additive.core.progress_logger import LOG
 from ansys.additive.core.simulation import SimulationError, SimulationStatus, SimulationType
 from ansys.additive.core.single_bead import (
     MeltPool,
